refactor: . move functions to EditableTask.ts
